requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
transformers==4.36.2
torch==2.1.2
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2 